backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
tensorflow==2.15.0
lightgbm==4.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
google-generativeai==0.3.2
nltk==3.8.1
feedparser==6.0.10
beautifulsoup4==4.12.2
schedule==1.2.0
psutil==5.9.6
cryptography==41.0.7
pytz==2023.3
websockets==12.0
discord.py==2.3.2
MetaTrader5==5.0.45
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.28.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
selenium==4.15.2
webdriver-manager==4.0.1
onnxruntime==1.16.3
joblib==1.3.2
ta==0.10.2